+++ FarmRearranger/docs/release-notes.md
@@ -2,6 +2,7 @@
 ## Upcoming release
 - The letter is now sent at the start of the day instead of overnight, so you get it immediately if you already meet the requirements.
 - Added in-game config UI through [Generic Mod Config Menu](https://www.nexusmods.com/stardewvalley/mods/5098).
+- Added support for split-screen mode.
 - Added failsafes to avoid the player getting stuck in invisible mode.
 - Fixed unable to activate farm rearranger when using a controller.